file uploader integrated with fileStorageAPI..

+++ services/src/main/java/si/fri/rso/services/FileManagerBean.java
@@ -4,6 +4,7 @@ import com.kumuluz.ee.discovery.annotations.DiscoverService;
 import org.apache.commons.io.FileUtils;
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import si.fri.rso.config.FileManagerConfigProperties;
+import si.fri.rso.lib.responses.ChannelBucketName;
 import si.fri.rso.lib.responses.NewFileMetadata;
 
 import javax.annotation.PostConstruct;
@@ -32,6 +33,7 @@ public class FileManagerBean {
     @DiscoverService(value = "rso1920-catalog")
     private Optional<String> fileMetadataUrl;
 
+
     private Client httpClient;
 
     @PostConstruct
@@ -48,55 +50,35 @@ public class FileManagerBean {
         } catch (IOException e) {
             e.printStackTrace();
         }
-        // TODO: call when rdy.
-        String filePath =  "new file path"; // requestSenderBean.sendFileToUploadOnS3(file);
-        //"c://bla//file.jpg"
 
         String [] nameType = fileDetails.getFileName().split("\\.");
-
 
         String fileType = nameType[nameType.length-1];
         Integer user = userChannel[0];
         Integer channel = userChannel[1];
 
+        ChannelBucketName channelBucketName = this.requestSenderBean.getBucketName(channel, requestUniqueID);
+
+        System.out.println("channel bucket name: " + channelBucketName.getBucketName());
+
+
+        boolean isFileUploaded = requestSenderBean.sendFileToUploadOnS3(file, channelBucketName.getBucketName(), fileDetails.getFileName(), requestUniqueID);
+
+        if (!isFileUploaded) {
+            System.out.println("NO UPLOADED");
+            return false;
+        }
+
         boolean isDeleted = file.delete();
         if(!isDeleted){
            System.out.println("File was not deleted!");
         }
+        String filePath =  channelBucketName.getBucketName() + "/" + fileDetails.getFileName(); //
+        NewFileMetadata newFile = new NewFileMetadata(filePath, fileDetails.getFileName(), fileType, user, channel);
 
-        if (filePath != null){
-            NewFileMetadata newFile = new NewFileMetadata(filePath, fileDetails.getFileName(), fileType, user, channel);
-            boolean isSaved = saveMetadata(newFile, requestUniqueID);
-            return true;
-        }
-        return false;
+        return requestSenderBean.saveMetadata(newFile, requestUniqueID);
     }
 
-    private boolean saveMetadata(NewFileMetadata newFile, String requestUniqueID){
-        if (!fileMetadataUrl.isPresent())
-            return false;
-        System.out.println("Saving metadata: " + fileMetadataUrl.get() + fileManagerConfigProperties.getCatalogApiUri());
-        System.out.println("REQUEST: " + requestUniqueID);
-        try{
-            Response success = httpClient
-                    .target(fileMetadataUrl.get() + fileManagerConfigProperties.getCatalogApiUri())
-                    .request(MediaType.APPLICATION_JSON_TYPE)
-                    .header("uniqueRequestId", requestUniqueID)
-                    .post( Entity.entity(newFile, MediaType.APPLICATION_JSON_TYPE));
-
-            if (success.getStatus() == 200) {
-                System.out.println("Meta data about file was uploaded");
-                return true;
-            } else {
-                System.out.println("Meta data about file was NOT uploaded");
-                return false;
-            }
-        }catch (WebApplicationException | ProcessingException e) {
-            // e.printStackTrace();
-            System.out.println("api for metadata upload failed miserably");
-            return false;
-        }
-    }
 
     //  1. upload zoro
     //  2. upload penca
@@ -105,12 +87,13 @@ public class FileManagerBean {
     public boolean deleteFile(Integer FileId, String path, String requestUniqueID) {
         String target = "";
         if (path.equals("storage")){
+            // TODO send request for file storage deletion..
             target = this.fileManagerConfigProperties.getFileStorageApiUri() + "/" + FileId;
         }
         else if (path.equals("catalog")){
             if (!fileMetadataUrl.isPresent())
                 return false;
-            target = fileMetadataUrl.get() + fileManagerConfigProperties.getDeletecatalogApiUri() +
+            target = fileMetadataUrl.get() + fileManagerConfigProperties.getDeleteCatalogApiUri() +
                     "/" +
                     FileId;
         }
@@ -138,5 +121,3 @@ public class FileManagerBean {
     }
 
 }
-
-// TODO maybe good idea to create one service specialized for calling other Microservices
